[eas-cli] remove unused classes, remove new directory
