+++ packages/eas-cli/src/credentials/android/actions/CreateFcm.ts
@@ -1,8 +1,8 @@
-import { AndroidFcmFragment, AndroidFcmVersion } from '../../../../graphql/generated';
-import Log from '../../../../log';
-import { promptAsync } from '../../../../prompts';
-import { Account } from '../../../../user/Account';
-import { Context } from '../../../context';
+import { AndroidFcmFragment, AndroidFcmVersion } from '../../../graphql/generated';
+import Log from '../../../log';
+import { promptAsync } from '../../../prompts';
+import { Account } from '../../../user/Account';
+import { Context } from '../../context';
 
 export class CreateFcm {
   constructor(private account: Account) {}
