+++ packages/eas-cli/src/credentials/android/actions/DownloadKeystore.ts
@@ -1,66 +1,82 @@
 import chalk from 'chalk';
 import fs from 'fs-extra';
 
+import { AndroidAppBuildCredentialsFragment } from '../../../graphql/generated';
 import Log from '../../../log';
 import { confirmAsync } from '../../../prompts';
 import { maybeRenameExistingFileAsync } from '../../../utils/files';
-import { Action, CredentialsManager } from '../../CredentialsManager';
 import { Context } from '../../context';
+import { AppLookupParams, formatProjectFullName } from '../api/GraphqlClient';
+import { displayAndroidKeystore } from '../utils/printCredentialsBeta';
 
 interface DownloadKeystoreOptions {
-  displayCredentials?: boolean;
+  app: AppLookupParams;
+  displaySensitiveInformation?: boolean;
   outputPath?: string;
 }
 
-export class DownloadKeystore implements Action {
-  constructor(private projectFullName: string, private options?: DownloadKeystoreOptions) {}
+export class DownloadKeystore {
+  constructor(private options: DownloadKeystoreOptions) {}
 
-  public async runAsync(manager: CredentialsManager, ctx: Context): Promise<void> {
-    const keystore = await ctx.android.fetchKeystoreAsync(this.projectFullName);
+  public async runAsync(
+    ctx: Context,
+    buildCredentials: AndroidAppBuildCredentialsFragment
+  ): Promise<void> {
+    const keystore = buildCredentials.androidKeystore;
     if (!keystore) {
-      Log.warn('There is no valid Keystore defined for this app');
+      Log.warn(
+        `There is no valid Keystore defined for build credentials: ${buildCredentials.name}`
+      );
       return;
     }
 
-    let displayCredentials = this.options?.displayCredentials;
-    if (displayCredentials === undefined && !ctx.nonInteractive) {
-      displayCredentials = await confirmAsync({
-        message: 'Do you want to display the Android Keystore credentials?',
+    let displaySensitiveInformation = this.options?.displaySensitiveInformation;
+    if (displaySensitiveInformation === undefined && !ctx.nonInteractive) {
+      displaySensitiveInformation = await confirmAsync({
+        message: 'Do you want to display the sensitive information of the Android Keystore?',
       });
     }
-
-    const keystorePath =
-      this.options?.outputPath ?? `${this.projectFullName.replace('/', '__')}.jks`;
+    const projectFullName = formatProjectFullName(this.options.app);
+    const keystorePath = this.options?.outputPath ?? `${projectFullName.replace('/', '__')}.jks`;
 
     await maybeRenameExistingFileAsync(ctx.projectDir, keystorePath);
 
     Log.log(`Saving Keystore to ${keystorePath}`);
     await fs.writeFile(keystorePath, keystore.keystore, 'base64');
 
-    if (displayCredentials) {
-      Log.log(`Keystore credentials
-  Keystore password: ${chalk.bold(keystore.keystorePassword)}
-  Key alias:         ${chalk.bold(keystore.keyAlias)}
-  Key password:      ${chalk.bold(keystore.keyPassword)}
+    // non-sensitive information
+    displayAndroidKeystore(keystore);
 
-  Path to Keystore:  ${keystorePath}
+    if (displaySensitiveInformation) {
+      Log.newLine();
+      Log.log(`Sensitive Keystore information:
+    Keystore password: ${chalk.bold(keystore.keystorePassword)}
+    Key alias:         ${chalk.bold(keystore.keyAlias)}
+    Key password:      ${chalk.bold(keystore.keyPassword)}
+
+    Path to Keystore:  ${keystorePath}
       `);
     }
+    Log.newLine();
   }
 }
 
-export class BackupKeystore implements Action {
-  constructor(private projectFullName: string) {}
+export class BackupKeystore {
+  constructor(private app: AppLookupParams) {}
 
-  public async runAsync(manager: CredentialsManager, ctx: Context): Promise<void> {
-    if (await ctx.android.fetchKeystoreAsync(this.projectFullName)) {
-      Log.warn('Backing up your old Android Keystore now...');
-      await manager.runActionAsync(
-        new DownloadKeystore(this.projectFullName, {
-          displayCredentials: true,
-          outputPath: `${this.projectFullName}.bak.jks`.replace('/', '__'),
-        })
-      );
+  public async runAsync(
+    ctx: Context,
+    buildCredentials: AndroidAppBuildCredentialsFragment
+  ): Promise<void> {
+    if (!buildCredentials.androidKeystore) {
+      return;
     }
+    const projectFullName = formatProjectFullName(this.app);
+    Log.warn('Backing up your old Android Keystore now...');
+    await new DownloadKeystore({
+      app: this.app,
+      outputPath: `${projectFullName}.bak.jks`.replace('/', '__'),
+      displaySensitiveInformation: true,
+    }).runAsync(ctx, buildCredentials);
   }
 }
