+++ packages/eas-cli/src/credentials/android/actions/__tests__/CreateFcm-test.ts
@@ -1,9 +1,9 @@
-import { promptAsync } from '../../../../../prompts';
-import { createCtxMock } from '../../../../__tests__/fixtures-context';
-import { getAppLookupParamsFromContext } from '../../BuildCredentialsUtils';
+import { promptAsync } from '../../../../prompts';
+import { createCtxMock } from '../../../__tests__/fixtures-context';
+import { getAppLookupParamsFromContext } from '../BuildCredentialsUtils';
 import { CreateFcm } from '../CreateFcm';
 
-jest.mock('../../../../../prompts');
+jest.mock('../../../../prompts');
 (promptAsync as jest.Mock).mockImplementation(() => ({ fcmApiKey: 'blah' }));
 
 describe(CreateFcm, () => {
