+++ packages/eas-cli/src/credentials/android/actions/__tests__/RemoveFcm-test.ts
@@ -1,13 +1,13 @@
-import { confirmAsync } from '../../../../../prompts';
+import { confirmAsync } from '../../../../prompts';
 import {
   getNewAndroidApiMockWithoutCredentials,
   testAndroidAppCredentialsFragment,
-} from '../../../../__tests__/fixtures-android-new';
-import { createCtxMock } from '../../../../__tests__/fixtures-context';
-import { getAppLookupParamsFromContext } from '../../BuildCredentialsUtils';
+} from '../../../__tests__/fixtures-android-new';
+import { createCtxMock } from '../../../__tests__/fixtures-context';
+import { getAppLookupParamsFromContext } from '../BuildCredentialsUtils';
 import { RemoveFcm } from '../RemoveFcm';
 
-jest.mock('../../../../../prompts');
+jest.mock('../../../../prompts');
 (confirmAsync as jest.Mock).mockImplementation(() => true);
 
 const originalConsoleLog = console.log;
